x86/uintr: Introduce user IPI sender syscalls
Add a registration syscall for a task to register itself as a user
interrupt sender using the uvec_fd generated by the receiver. A task
can register multiple uvec_fds. Each unique successful connection
creates a new entry in the User Interrupt Target Table (UITT).

Each entry in the UITT table is referred by the UITT index (uipi_index).
The uipi_index returned during the registration syscall lets a sender
generate a user IPI using the 'SENDUIPI <uipi_index>' instruction.

Also, add a sender unregister syscall to unregister a particular task
from the uvec_fd. Calling close on the uvec_fd will disconnect all
threads in a sender process from that FD.

Currently, the UITT size is arbitrarily chosen as 256 entries
corresponding to a 4KB page. Based on feedback and usage data this can
either be increased/decreased or made dynamic later.

Architecturally, the UITT table can be unique for each thread or shared
across threads of the same thread group. The benefit of sharing the UITT
table is that all threads would see the same view of the UITT table.
Also the kernel UITT memory allocation would be more efficient if
multiple threads connect to the same uvec_fd. However, this would mean
the kernel needs to keep the UITT table size MISC_MSR[] in sync across
these threads. Also the UPID/UITT teardown flows might need additional
consideration.

Another option is to keep the UITT as unique for the each thread. This
makes the kernel implemantion relatively simple and only threads that
use uintr get setup with the related structures. However, this means
that the uipi_index for each thread would be inconsistent wrt to other
threads.  (Executing 'SENDUIPI 2' on threads of the same process could
generate different user interrupts.)

The current implementation shares the UITT for tasks of the same process
address space.  A #UD is generated when a task tries to use the UITT
without it being programmed by the kernel. Use that #UD to fix and patch
the task MSRs.

Signed-off-by: Sohil Mehta <sohil.mehta@intel.com>